backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==14.1
httpx==0.28.1
python-dotenv==1.0.1
aiohttp==3.11.10
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.0
numpy==2.2.1
pandas==2.2.3
scipy==1.15.0
py-clob-client==0.28.0
requests==2.32.3
pytz==2024.2
python-dateutil==2.9.0
openai==1.59.3
